refactored and renamed UpperTabs to Tabs
